Refactoring week5 LiveCoding sessions Replaced auto with actual types. Added IsValid checks and logs. Fixed Haste skill. Renamed teleport to Teleport.
