+++ Content/TopDownCPP/Blueprints/Skills/Haste.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fd60fe6ca1ff296b9ef4b925c9e76cd6cdbf870a06622ab9fdc4c4ea21369e14
-size 16618
+oid sha256:14557751b165208fc012f821c760b8789e20b270074d9f397be7c9091847925f
+size 26808
