+++ Content/TopDownCPP/Blueprints/Skills/teleport.uasset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:409e80b7a01313bfea0e6c5fa58c184337e821287759be21a3ceb5bead53a4b7
-size 29155
+oid sha256:ae8cc77a6274f07c8f4c61b47a1a6eba5ac50add7e744b4824009ff70ec31091
+size 27788
