+++ Source/TopDownARPG/Abilities/Ability.cpp
@@ -2,11 +2,16 @@
 
 
 #include "Ability.h"
+#include "TopDownARPG.h"
 
 void UAbility::Activate(AActor* Source)
 {
-    UWorld* World = Source->GetWorld();
-    check(World);
+    UWorld* World = GetWorld();
+	if (IsValid(World) == false)
+	{
+		UE_LOG(LogTopDownARPG, Error, TEXT("UAbility::Activate IsValid(World) == false"));
+		return;
+	}
     TimerManager = &World->GetTimerManager();
     TimerManager->SetTimer(CooldownTimerHandle, this, &UAbility::OnCooldownTimerExpired, CooldownTimeDilation, true, CooldownTime);
     bIsOffCooldown = false;
